function/nats-function/nats-function/requirements.txt
nats-py
requests